+++ edurange_refactored/flask/modules/utils/auth_utils.py
@@ -4,10 +4,9 @@ from flask import (
     jsonify,
     g
 )
-from edurange_refactored.user.models import User
+
 from functools import wraps
 from flask_jwt_simple import decode_jwt
-from edurange_refactored.user.models import GroupUsers, StudentGroups, User
 
 
 ###########
@@ -56,19 +55,3 @@ def jwt_and_csrf_required(fn):
         return fn(*args, **kwargs)
     
     return wrapper
-
-def register_user(validated_registration_data):
-    data = validated_registration_data
-    print("trying to register_user() with: ", data)
-    User.create(
-            username=data["username"],
-            email=data["email"],
-            password=data["password"],
-            active=True,
-    )
-       
-    group = StudentGroups.query.filter_by(code=data["code"]).first()
-    user = User.query.filter_by(username=data["username"]).first()
-    group_id = group.get_id()
-    user_id = user.get_id()
-    GroupUsers.create(user_id=user_id, group_id=group_id)
